feat: Milestone 2+3 - settings blueprint + wrapper reduction sweep
Agent-Logs-Url: https://github.com/abartrim/sobs/sessions/30b0c283-47d6-4ae8-9f23-779024f9e70b
