+++ templates/ai.html
@@ -229,7 +229,7 @@
         <a
           id="estCostInferredHint"
           class="d-none text-info"
-          href="{{ url_for('view_ai_settings') }}"
+          href="{{ url_for('settings.view_ai_settings') }}"
           title="Includes inferred pricing"
           aria-label="Includes inferred pricing"
         ><i class="bi bi-info-circle"></i></a>
