+++ templates/ai_help.html
@@ -132,10 +132,10 @@
         <div class="card-body">
             <ul class="small mb-0">
                 <li>
-                    <a href="{{ url_for('view_ai_settings') }}">AI Settings</a> — configure LLM provider, model, and guard settings.
+                    <a href="{{ url_for('settings.view_ai_settings') }}">AI Settings</a> — configure LLM provider, model, and guard settings.
                 </li>
                 <li>
-                    <a href="{{ url_for('view_agent_rules') }}">Agent Settings</a> — create and manage autonomous agent rules.
+                    <a href="{{ url_for('settings.view_agent_rules') }}">Agent Settings</a> — create and manage autonomous agent rules.
                 </li>
                 <li>
                     <a href="{{ url_for('view_work_items') }}">Work Items</a> — GitHub issues created by agent runs.
