+++ templates/cve.html
@@ -27,7 +27,7 @@
 <div class="alert alert-secondary py-2 small" role="alert">
   <i class="bi bi-shield-slash me-1"></i>
   CVE scanning is <strong>disabled</strong>.
-  <a href="{{ url_for('view_enrichment_settings') }}" class="alert-link">Enable in Enrichment Settings</a>
+  <a href="{{ url_for('settings.view_enrichment_settings') }}" class="alert-link">Enable in Enrichment Settings</a>
 </div>
 {% else %}
