+++ templates/settings_help.html
@@ -68,11 +68,11 @@
             <div class="row g-2">
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-primary w-100"
-                       href="{{ url_for('view_ai_settings') }}">AI Settings</a>
+                       href="{{ url_for('settings.view_ai_settings') }}">AI Settings</a>
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-warning w-100"
-                       href="{{ url_for('view_agent_rules') }}">Agent Rules</a>
+                       href="{{ url_for('settings.view_agent_rules') }}">Agent Rules</a>
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-warning w-100"
@@ -88,7 +88,7 @@
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-secondary w-100"
-                       href="{{ url_for('view_settings_repositories') }}">Repositories</a>
+                       href="{{ url_for('settings.view_settings_repositories') }}">Repositories</a>
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-info w-100"
