+++ templates/summary.html
@@ -310,7 +310,7 @@
         {% else %}
         <div class="text-secondary small">
           CVE scanning is disabled.
-          <a href="{{ url_for('view_enrichment_settings') }}">Enable in enrichment settings</a>.
+          <a href="{{ url_for('settings.view_enrichment_settings') }}">Enable in enrichment settings</a>.
         </div>
         {% endif %}
       </div>
