+++ templates/web_traffic.html
@@ -66,7 +66,7 @@
 <div class="alert alert-secondary py-2 small mb-3" role="alert">
   <i class="bi bi-geo-alt-fill me-1"></i>
   Geo lookup is <strong>disabled</strong>.
-  <a href="{{ url_for('view_enrichment_settings') }}" class="alert-link">Enable in settings</a>
+  <a href="{{ url_for('settings.view_enrichment_settings') }}" class="alert-link">Enable in settings</a>
 </div>
 {% endif %}
